+++ main.go
@@ -12,10 +12,10 @@ import (
 
 func init() {
 	log.Println("IGIS IDP Platform Backend Starting at http://127.0.0.1:8080")
-	log.Println("RowCountTest: ", apis.TEST_URL_ROW)
 	log.Println("AssetTest: ", apis.TEST_URL_ASSET)
 	log.Println("DebtTest: ", apis.TEST_URL_DEBT)
 	log.Println("MacroTest: ", apis.TEST_URL_MACRO)
+	log.Println("SingleTest: ", apis.TEST_URL_SINGLE)
 }
 
 func main() {
@@ -34,19 +34,30 @@ func main() {
 		HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			apis.ServeDebt(d.Debt, 0, w, r)
 		}).
+		Methods("GET").
 		Name("debt.datatable")
 	sV1Debt.Path("/graphLeft").
 		HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			apis.ServeDebt(d.Debt, 1, w, r)
 		}).
-		Name("debt.graph")
+		Methods("GET").
+		Name("debt.graphLeft")
 	sV1Debt.Path("/graphRight").
 		HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			apis.ServeDebt(d.Debt, 2, w, r)
-		})
+		}).
+		Methods("GET").
+		Name("debt.graphRight")
 
 	// api v1 subrouter all
 	sV1 := r.PathPrefix("/api/v1").Subrouter()
+	sV1.HandleFunc("/single", func(w http.ResponseWriter, r *http.Request) {
+		apis.ServeSingle(d.Debt, w, r)
+	}).
+		Methods("GET").
+		Name("singleInfo")
+
+	// TODO: EDIT Below
 	sV1.HandleFunc("/asset", func(w http.ResponseWriter, r *http.Request) {
 		apis.ServeAssetWhole(d.Asset, w, r)
 	}).